Moved LOD test into test_plot
